WIP. move some common methods to parent class
